refactor python tests and dependencies
