+++ python/tests/test_auth.py
