+++ python/tests/test_base_install/test_graphql/edit_graph/test_delete_graph.py
