+++ python/tests/test_base_install/test_graphql/misc/test_snapshot.py
