+++ python/tests/test_base_install/test_graphql/test_namespace.py
