+++ python/tests/test_base_install/test_vectors.py
