+++ python/tox.ini
@@ -1,0 +1,34 @@
+[tox]
+requires =
+    tox>=4
+env_list = base, export, all
+
+# MATURIN_PEP517_ARGS
+[testenv]
+description = run the tests with pytest
+package = wheel
+wheel_build_env = .pkg
+deps =
+    pytest>=8
+pass_env =
+    DISK_TEST_MARK
+
+[testenv:.pkg]
+pass_env = MATURIN_PEP517_ARGS
+
+[testenv:export]
+extras =
+    tox
+    export
+commands = pytest --nbmake --nbmake-timeout=1200 {tty:--color=yes} tests/test_export.py
+
+[testenv:base]
+extras =
+    tox
+commands = pytest --nbmake --nbmake-timeout=1200 {tty:--color=yes} tests/test_base_install
+
+[env:all]
+extras =
+    tox
+    test
+commands = pytest --nbmake --nbmake-timeout=1200 {tty:--color=yes} tests
